+++ rdt/transformers/boolean.py
@@ -43,7 +43,7 @@ class BinaryEncoder(BaseTransformer):
                  missing_value_generation='random'):
         super().__init__()
         self._set_missing_value_generation(missing_value_generation)
-        self._set_missing_value_replacement('mode', missing_value_replacement)
+        self._set_missing_value_replacement('random', missing_value_replacement)
         if model_missing_values is not None:
             self._set_model_missing_values(model_missing_values)