+++ tests/unit/transformers/test_categorical.py
@@ -344,6 +344,19 @@ class TestOrderedUniformEncoder:
         # Asserts
         pd.testing.assert_series_equal(transformer.order, pd.Series(['b', 'c', 'a', np.nan]))
 
+    def test___init___duplicate_categories(self):
+        """Test the ``__init__`` method errors if duplicate categories provided.
+
+        Test initialization errors if duplicate categories are passed in the ``order`` parameter.
+        """
+        # Run / Assert
+        expected_msg = (
+            "The OrderedUniformEncoder has duplicate categories in the 'order' parameter. "
+            'Please drop the duplicates to proceed.'
+        )
+        with pytest.raises(TransformerInputError, match=expected_msg):
+            OrderedUniformEncoder(order=['a', 'b', 'c', 'c'])
+
     def test___repr___default(self):
         """Test that the ``__repr__`` method prints the custom order.
 
@@ -512,7 +525,12 @@ class TestFrequencyEncoder:
     def test___init__(self):
         """Passed arguments must be stored as attributes."""
         # Run
-        transformer = FrequencyEncoder(add_noise='add_noise_value')
+        warn_message = (
+            "The 'FrequencyEncoder' transformer will no longer be supported in future "
+            "versions of the RDT library. Please use the 'UniformEncoder' transformer instead."
+        )
+        with pytest.warns(FutureWarning, match=warn_message):
+            transformer = FrequencyEncoder(add_noise='add_noise_value')
 
         # Asserts
         assert transformer.add_noise == 'add_noise_value'
@@ -700,11 +718,11 @@ class TestFrequencyEncoder:
         # Asserts
         assert result == 0.2745
 
-    def test__reverse_transform_array(self):
-        """Test reverse_transform a numpy.array"""
+    def test__reverse_transform_series(self):
+        """Test reverse_transform a pandas Series"""
         # Setup
         data = pd.Series(['foo', 'bar', 'bar', 'foo', 'foo', 'tar'])
-        rt_data = np.array([-0.6, 0.5, 0.6, 0.2, 0.1, -0.2])
+        rt_data = pd.Series([-0.6, 0.5, 0.6, 0.2, 0.1, -0.2])
         transformer = FrequencyEncoder()
 
         # Run
@@ -987,83 +1005,14 @@ class TestFrequencyEncoder:
         expected = np.array([0.875, 0.625, 0.375, 0.125])
         assert (transformed == expected).all()
 
-    @patch('psutil.virtual_memory')
-    def test__reverse_transform_by_matrix_called(self, psutil_mock):
-        """Test that the `_reverse_transform_by_matrix` method is called.
-
-        When there is enough virtual memory, expect that the
-        `_reverse_transform_by_matrix` method is called.
-
-        Setup:
-            The categorical transformer is instantiated with 4 categories. Also patch the
-            `psutil.virtual_memory` function to return a large enough `available_memory`.
-        Input:
-            - numerical data with 4 rows
-        Output:
-            - the output of `_reverse_transform_by_matrix`
-        Side effects:
-            - `_reverse_transform_by_matrix` will be called once
-        """
-        # Setup
-        data = pd.Series([1, 2, 3, 4])
-
-        categorical_transformer_mock = Mock()
-        categorical_transformer_mock.means = pd.Series([0.125, 0.375, 0.625, 0.875])
-
-        virtual_memory = Mock()
-        virtual_memory.available = 4 * 4 * 8 * 3 + 1
-        psutil_mock.return_value = virtual_memory
-
-        # Run
-        reverse = FrequencyEncoder._reverse_transform(categorical_transformer_mock, data)
-
-        # Asserts
-        reverse_arg = categorical_transformer_mock._reverse_transform_by_matrix.call_args[0][0]
-        np.testing.assert_array_equal(reverse_arg, data.clip(0, 1))
-        assert reverse == categorical_transformer_mock._reverse_transform_by_matrix.return_value
-
-    @patch('psutil.virtual_memory')
-    def test__reverse_transform_by_matrix(self, psutil_mock):
-        """Test the _reverse_transform_by_matrix method with numerical data.
-
-        Expect that the transformed data is correctly reverse transformed.
-
-        Setup:
-            The categorical transformer is instantiated with 4 categories and means. Also patch
-            the ``psutil.virtual_memory`` function to return a large enough ``available_memory``.
-        Input:
-            - transformed data with 4 rows
-        Ouptut:
-            - the original data
-        """
-        # Setup
-        data = pd.Series([1, 2, 3, 4])
-        transformed = pd.Series([0.875, 0.625, 0.375, 0.125])
-
-        transformer = FrequencyEncoder()
-        transformer.starts = pd.DataFrame({'category': [4, 3, 2, 1]}, index=[0., 0.25, 0.5, 0.75])
-        transformer.dtype = data.dtype
-
-        virtual_memory = Mock()
-        virtual_memory.available = 4 * 4 * 8 * 3 + 1
-        psutil_mock.return_value = virtual_memory
-
-        # Run
-        reverse = transformer._reverse_transform_by_matrix(transformed)
-
-        # Assert
-        pd.testing.assert_series_equal(data, reverse)
-
-    @patch('psutil.virtual_memory')
-    def test__reverse_transform_by_category_called(self, psutil_mock):
+    def test__reverse_transform_by_category_called(self):
         """Test that the `_reverse_transform_by_category` method is called.
 
-        When there is not enough virtual memory and the number of rows is greater than the
-        number of categories, expect that the `_reverse_transform_by_category` method is called.
+        When the number of rows is greater than the number of categories, expect
+        that the `_reverse_transform_by_category` method is called.
 
         Setup:
-            The categorical transformer is instantiated with 4 categories. Also patch the
-            `psutil.virtual_memory` function to return an `available_memory` of 1.
+            The categorical transformer is instantiated with 4 categories.
         Input:
             - numerical data with 5 rows
         Output:
@@ -1077,10 +1026,6 @@ class TestFrequencyEncoder:
         categorical_transformer_mock = Mock()
         categorical_transformer_mock.means = pd.Series([0.125, 0.375, 0.625, 0.875])
 
-        virtual_memory = Mock()
-        virtual_memory.available = 1
-        psutil_mock.return_value = virtual_memory
-
         # Run
         reverse = FrequencyEncoder._reverse_transform(
             categorical_transformer_mock, transform_data)
@@ -1090,16 +1035,14 @@ class TestFrequencyEncoder:
         np.testing.assert_array_equal(reverse_arg, transform_data.clip(0, 1))
         assert reverse == categorical_transformer_mock._reverse_transform_by_category.return_value
 
-    @patch('psutil.virtual_memory')
-    def test__reverse_transform_by_category(self, psutil_mock):
+    def test__reverse_transform_by_category(self):
         """Test the _reverse_transform_by_category method with numerical data.
 
         Expect that the transformed data is correctly reverse transformed.
 
         Setup:
             The categorical transformer is instantiated with 4 categories, and the means
-            and intervals are set for those categories. Also patch the `psutil.virtual_memory`
-            function to return an `available_memory` of 1.
+            and intervals are set for those categories.
         Input:
             - transformed data with 5 rows
         Ouptut:
@@ -1117,10 +1060,6 @@ class TestFrequencyEncoder:
             1: (0.75, 1.0, 0.875, 0.041666666666666664),
         }
         transformer.dtype = data.dtype
-
-        virtual_memory = Mock()
-        virtual_memory.available = 1
-        psutil_mock.return_value = virtual_memory
 
         reverse = transformer._reverse_transform_by_category(transformed)
 
@@ -1152,17 +1091,14 @@ class TestFrequencyEncoder:
         # Assert
         assert category == 'c'
 
-    @patch('psutil.virtual_memory')
-    def test__reverse_transform_by_row_called(self, psutil_mock):
+    def test__reverse_transform_by_row_called(self):
         """Test that the `_reverse_transform_by_row` method is called.
 
-        When there is not enough virtual memory and the number of rows is less than or equal
-        to the number of categories, expect that the `_reverse_transform_by_row` method
-        is called.
+        When the number of rows is less than or equal to the number of categories,
+        expect that the `_reverse_transform_by_row` method is called.
 
         Setup:
-            The categorical transformer is instantiated with 4 categories. Also patch the
-            `psutil.virtual_memory` function to return an `available_memory` of 1.
+            The categorical transformer is instantiated with 4 categories.
         Input:
             - numerical data with 4 rows
         Output:
@@ -1179,10 +1115,6 @@ class TestFrequencyEncoder:
             [0., 0.25, 0.5, 0.75], index=[4, 3, 2, 1], columns=['category'])
         categorical_transformer_mock._normalize.return_value = data
 
-        virtual_memory = Mock()
-        virtual_memory.available = 1
-        psutil_mock.return_value = virtual_memory
-
         # Run
         reverse = FrequencyEncoder._reverse_transform(categorical_transformer_mock, data)
 
@@ -1191,16 +1123,14 @@ class TestFrequencyEncoder:
         np.testing.assert_array_equal(reverse_arg, data.clip(0, 1))
         assert reverse == categorical_transformer_mock._reverse_transform_by_row.return_value
 
-    @patch('psutil.virtual_memory')
-    def test__reverse_transform_by_row(self, psutil_mock):
+    def test__reverse_transform_by_row(self):
         """Test the _reverse_transform_by_row method with numerical data.
 
         Expect that the transformed data is correctly reverse transformed.
 
         Setup:
             The categorical transformer is instantiated with 4 categories, and the means, starts,
-            and intervals are set for those categories. Also patch the `psutil.virtual_memory`
-            function to return an `available_memory` of 1.
+            and intervals are set for those categories.
         Input:
             - transformed data with 4 rows
         Ouptut:
@@ -1221,10 +1151,6 @@ class TestFrequencyEncoder:
             1: (0.75, 1.0, 0.875, 0.041666666666666664),
         }
         transformer.dtype = data.dtype
-
-        virtual_memory = Mock()
-        virtual_memory.available = 1
-        psutil_mock.return_value = virtual_memory
 
         # Run
         reverse = transformer._reverse_transform(transformed)
@@ -2278,6 +2204,19 @@ class TestOrderedLabelEncoder:
         # Asserts
         assert transformer.add_noise == 'add_noise_value'
         pd.testing.assert_series_equal(transformer.order, pd.Series(['b', 'c', 'a', np.nan]))
+
+    def test___init___duplicate_categories(self):
+        """The the ``__init__`` method with duplicate categories in the order parameter.
+
+        Transformer should error with ``TransformerInputError``.
+        """
+        # Run / Assert
+        expected_msg = (
+            "The OrderedLabelEncoder has duplicate categories in the 'order' parameter. "
+            'Please drop the duplicates to proceed.'
+        )
+        with pytest.raises(TransformerInputError, match=expected_msg):
+            OrderedLabelEncoder(order=['b', 'c', 'a', 'a'], add_noise='add_noise_value')
 
     def test___repr___default(self):
         """Test that the ``__repr__`` method prints the custom order.